refactor(consensus): Pass D DEF8/DEF9 — float→uint32 probability proto fields
FighterName.Probability (field 5), FighterType.Probability (field 3) and
FighterTypeMoveProbability.Probability (field 2) were `float` but every config
value is an exact non-negative integer (FighterName 1..1000; FighterType 20;
move probs 15/20/30). Reading them as float into the recipe/fighter generation
RNG (NextInt bounds + candidate thresholds) is a determinism-discipline risk in
consensus code. Change the three fields to `uint32`, rewrite fightertype.pb.text
`Probability: N.0` → `N`, and give recipe.cpp's six `*1000` sites explicit
`(int32_t)` casts so the candidate math is pure signed integer.

Golden byte-identical (all values exact integers → identical NextInt bounds and
comparisons), 98 unit + 4 reorg + 2 reorg-game + proto2 roconfig + database all
green. No set_probability writer exists (config-read-only); roconfig_tests only
check has_probability() → uint32-safe. Full clean rebuild per the build gotcha.

A read-only 4-lens determinism completeness sweep run alongside this change
re-verified DEF8/DEF9 value-exactness and surfaced 4 more float-in-consensus
proto fields the original audit missed (DEF10-14, tracked in the findings doc);
DEF10/11 are golden-neutral, DEF12-14 (fractional, incl. the exchange money
path) await a representation decision.

+++ database/recipe.cpp
@@ -327,9 +327,9 @@ uint32_t RecipeInstance::Generate(pxd::Quality quality, const RoConfig& cfg,  xa
      
     for (const auto& fighter : sortedfighterTypesmap)
     {
-        if(probabilityTreshholdMaximum <  fighter.second.probability() * 1000)
+        if(probabilityTreshholdMaximum <  (int32_t) fighter.second.probability() * 1000)
         {
-            probabilityTreshholdMaximum = fighter.second.probability() * 1000;
+            probabilityTreshholdMaximum = (int32_t) fighter.second.probability() * 1000;
         }
     }
         
@@ -338,7 +338,7 @@ uint32_t RecipeInstance::Generate(pxd::Quality quality, const RoConfig& cfg,  xa
     {
       for (const auto& fighter : sortedfighterTypesmap)
       {
-        int32_t probabilityTreshhold = fighter.second.probability() * 1000;
+        int32_t probabilityTreshhold = (int32_t) fighter.second.probability() * 1000;
         int32_t rolCurNum = rnd.NextInt(probabilityTreshholdMaximum * 2);
 
         if(rolCurNum < probabilityTreshhold)
@@ -436,9 +436,9 @@ uint32_t RecipeInstance::Generate(pxd::Quality quality, const RoConfig& cfg,  xa
      
     for(auto& probableMove: fighterType.moveprobabilities())
     {
-        if(probabilityTreshholdMaximum <  probableMove.probability() * 1000)
+        if(probabilityTreshholdMaximum <  (int32_t) probableMove.probability() * 1000)
         {
-            probabilityTreshholdMaximum = probableMove.probability() * 1000;
+            probabilityTreshholdMaximum = (int32_t) probableMove.probability() * 1000;
         }
     }    
     
@@ -450,7 +450,7 @@ uint32_t RecipeInstance::Generate(pxd::Quality quality, const RoConfig& cfg,  xa
             
             for(auto& probableMove: fighterType.moveprobabilities())
             {
-                int32_t probabilityTreshhold = probableMove.probability() * 1000;
+                int32_t probabilityTreshhold = (int32_t) probableMove.probability() * 1000;
                 
                 int32_t rolCurNum = rnd.NextInt(probabilityTreshholdMaximum * 2);  
                 if(rolCurNum < probabilityTreshhold)
